add all utils to util.js
