--- js/utils/get-random-integer.js
@@ -1,12 +1,0 @@
-const getRandomInteger = (min, max) => {
-  if (min < 0 || max < 0) {
-    return Error('Arguments must be positive ("min" number can equal 0)');
-  }
-  if (min >= max) {
-    return Error('The "max" number must be greater than "min" number');
-  }
-
-  return Math.floor(Math.random() * (Math.ceil(max) - Math.ceil(min) + 1)) + Math.ceil(min);
-};
-
-export { getRandomInteger };
